requirements.txt
Django==1.4.5
psycopg2==2.4.6
Pillow==2.0.0
gunicorn==0.17.2
django-redis-cache==0.9.7
south==0.7.6
dj-database-url==0.2.1
django-debug-toolbar==0.9.4
sqlalchemy==0.8.0 #for connection pooling
newrelic==1.10.2.38 #heroku performance monitoring

#storage
-e hg+https://bitbucket.org/david/django-storages@561f3dd076537b83d595bf34898f6961ba8f38b3#egg=django-storages
boto==2.8.0
django-compressor==1.3
django-cached-s3-storage==0.2

-e git://github.com/divio/django-cms.git@master#egg=django-cms
django-cms-themes==1.0.10
django-tinymce==1.5.1b4